+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-- Cleanup
+- Improved Custom Trim Combinations Screen
+  - Previews of the trims will now change accordingly to the material to apply on
+  - Previews of not valid options will now be displayed as a barriers
+- Fixed leather not being valid for materials to apply custom trim combinations on
 - Rewritten config backend